chore: add some docs, remove invalid constant
