+++ telegram/message/inline/types.go
@@ -1,20 +1,45 @@
 package inline
 
 const (
-	PhotoType   = "photo"
+	// PhotoType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultphoto.
+	PhotoType = "photo"
+	// ArticleType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultarticle.
 	ArticleType = "article"
 
-	VideoType     = "video"
-	AnimationType = "animation"
-	AudioType     = "audio"
-	DocumentType  = "document"
-	GIFType       = "gif"
-	MPEG4GIFType  = "mpeg4_gif"
-	VoiceType     = "voice"
-	StickerType   = "sticker"
+	// VideoType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultvideo.
+	VideoType = "video"
+	// AudioType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultaudio.
+	AudioType = "audio"
+	// DocumentType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultdocument.
+	DocumentType = "document"
+	// GIFType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultgif.
+	GIFType = "gif"
+	// MPEG4GIFType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultmpeg4gif.
+	MPEG4GIFType = "mpeg4_gif"
+	// VoiceType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultvoice.
+	VoiceType = "voice"
+	// StickerType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultsticker.
+	StickerType = "sticker"
 
+	// LocationType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultlocation.
 	LocationType = "location"
-	VenueType    = "venue"
-	ContactType  = "contact"
-	GameType     = "game"
+	// VenueType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultvenue.
+	VenueType = "venue"
+	// ContactType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultcontact.
+	ContactType = "contact"
+	// GameType is a type string of inline result.
+	// See https://core.telegram.org/bots/api#inlinequeryresultgame.
+	GameType = "game"
 )
